Move more work into encode_timestamp and decode_timestamp helpers.
